fixing page -routing
